Update README.md
It explains why I am embarked on such a seemingly bizarre task, to find the programs lurking in our "bin" directories that are not compiled programs (ELF in Linux, different in Windows - I use Windows 10), but scripts that you can copy to your own directories and read directly. It is instructional, interesting, and just plain fun to do this.