This is to extend the program to analyze "bin" dirs
The old program only analyzed the scripts that could be found in /usr/bin. This branch aims to analyze any of the numerous "bin" directories that have compiled programs as well as scripts.

I also have decided to use classes, two in particular. One will hold the raw data of the directory, the second "scripts" will hold the information about scripts that are found among the compiled programs, and what their outstanding traits are (language, size, name of program, etc).