I decided that, among other improvements, it might make things neater to use classes.
"get_scripts_with_a_class.py  is the next phase of the program that demonstrates the idea (The old program is "get_scripts_in_usr_bin.py", in the "master" branch. I have a long way to go to finish this program, and the shell script that runs the program, (temporarily, s2.sh). In any case, comments are welcomed and collaborators eagerly sought.