The entry point of this  system is the program "scripts-search.py"
Scripts-search is starting to work on an elementary level. Not only will you be able to cull "scripts" out of a directory of executables; they can also be sorted by program name, program type (shell script, Ruby script, Lua, etc) or size. It has not made its way into the command-line arguments (I am still somewhat mystified by "argparse"). But the machinery is there.

My goal is to have a slick, good-looking report, or possibly series of spreadsheets, and a report that breaks down the statistics of the directory, as far as scripts are concerned.
To this end, I am looking into using python/GTK+ as a front-end.