Outputs .CSV files - for spreadsheets.
This program now puts out data in .csv format, with titles on the first line. So far it has imported well into LibreOffice and MS-Excel (caution: this is a Linux/Unix program). There arew four fields (columns): 1) the name of the program 2) A brief "whatis" summary of the program, when available, 3) The type of the program (e.g. Bourne Shell script, Python Script, Ruby, Perl, Lua, REXX, makefile ro symlink),, 4) Size, in bytes, of the program.